Changed "Basics" text
Changed the teaser text for "Basics" from "You probably already use stories and reason about forces to come up with good solutions. Embodied Making gives structure to make your personal process shareable with others. Learn the basics to learn more." to "There are probably some things in life you do with embodiment. Like riding a bicycle or playing a sport. It usually takes years of personal experience before you can do things with embodiment. And it usually remains personal. It's easier to feel embodiment than explain it. And it's easy enough to spot people doing things in a disembodied way when you're doing it in an embodied way. Embodied Making is a method to help you develop embodiment together with others for a shared deep feeling and context."